+++ crates/runner-host/src/board/platform.rs
@@ -17,46 +17,29 @@ use std::borrow::Cow;
 use data_encoding::HEXLOWER_PERMISSIVE;
 use wasefire_board_api::platform::Api;
 use wasefire_board_api::Error;
-use wasefire_error::Code;
+
+use crate::FLAGS;
 
 pub mod protocol;
+mod update;
 
 pub enum Impl {}
 
 impl Api for Impl {
     type Protocol = protocol::Impl;
 
-    type Update = UpdateImpl;
+    type Update = update::Impl;
 
     fn serial() -> Cow<'static, [u8]> {
-        from_hex(option_env!("WASEFIRE_HOST_SERIAL"))
+        from_hex(FLAGS.serial.as_deref())
     }
 
     fn version() -> Cow<'static, [u8]> {
-        from_hex(option_env!("WASEFIRE_HOST_VERSION"))
+        from_hex(FLAGS.version.as_deref())
     }
 
     fn reboot() -> Result<!, Error> {
-        Err(Error::world(Code::NotImplemented))
-    }
-}
-
-pub enum UpdateImpl {}
-impl wasefire_board_api::Support<bool> for UpdateImpl {
-    const SUPPORT: bool = false;
-}
-impl wasefire_board_api::platform::update::Api for UpdateImpl {
-    fn metadata() -> Result<Box<[u8]>, Error> {
-        Err(Error::world(Code::NotImplemented))
-    }
-    fn initialize(_dry_run: bool) -> Result<(), Error> {
-        Err(Error::world(Code::NotImplemented))
-    }
-    fn process(_chunk: &[u8]) -> Result<(), Error> {
-        Err(Error::world(Code::NotImplemented))
-    }
-    fn finalize() -> Result<(), Error> {
-        Err(Error::world(Code::NotImplemented))
+        crate::cleanup::shutdown(0)
     }
 }