+++ crates/xtask/src/main.rs
@@ -190,6 +190,12 @@ struct RunnerOptions {
     #[clap(long)]
     version: Option<String>,
 
+    /// Host platform serial.
+    ///
+    /// This is only used for the host runner. It must be an hexadecimal byte sequence.
+    #[clap(long)]
+    serial: Option<String>,
+
     /// Cargo no-default-features.
     #[clap(long)]
     no_default_features: bool,
@@ -465,7 +471,12 @@ impl RunnerOptions {
         if self.name == "host" {
             if let Some(version) = &self.version {
                 cargo.env("WASEFIRE_HOST_VERSION", version);
-            };
+            }
+            if let Some(serial) = &self.serial {
+                cargo.env("WASEFIRE_HOST_SERIAL", serial);
+            }
+            cmd::execute(Command::new("make").current_dir("crates/runner-host/crates/web-client"))
+                .await?;
         }
         if self.name == "nordic" {
             rustflags.push(format!("-C link-arg=--defsym=RUNNER_SIDE={step}"));
@@ -533,7 +544,7 @@ impl RunnerOptions {
                     }
                 }
             }
-            cargo.arg("--");
+            cargo.args(["--", "../../target/wasefire"]);
             if std::env::var_os("CODESPACES").is_some() {
                 log::warn!("Assuming runner --arg=--protocol=unix when running in a codespace.");
                 cargo.arg("--protocol=unix");
@@ -600,10 +611,14 @@ impl RunnerOptions {
             }
         }
         if matches!(cmd, Some(RunnerCommand::Bundle)) {
-            let mut objcopy = wrap_command().await?;
-            objcopy.args(["rust-objcopy", "-O", "binary", &elf]);
-            objcopy.arg(format!("target/wasefire/platform{side}.bin"));
-            cmd::execute(&mut objcopy).await?;
+            let bundle = format!("target/wasefire/platform{side}.bin");
+            if self.name == "host" {
+                fs::copy(elf, bundle).await?;
+            } else {
+                let mut objcopy = wrap_command().await?;
+                objcopy.args(["rust-objcopy", "-O", "binary", &elf, &bundle]);
+                cmd::execute(&mut objcopy).await?;
+            }
             if step < max_step {
                 return Box::pin(self.execute(main, step + 1, cmd)).await;
             }